fix cast

+++ optee_qemu/hello_ianni/host/main.c
@@ -81,7 +81,7 @@ int main(int argc, char *argv[])
 	 */
 	op.paramTypes = TEEC_PARAM_TYPES(TEEC_VALUE_INOUT, TEEC_NONE,
 					 TEEC_NONE, TEEC_NONE);
-	op.params[0].value.a = test_prefetch_abort;
+	op.params[0].value.a = (unsigned long)test_prefetch_abort;
 
 	/*
 	 * TA_HELLO_WORLD_CMD_INC_VALUE is the actual function in the TA to be
